Rudimentary Chess AI
This project was programmed as part of a group of three and was
completed January 2012. This was a final project of a high school
introductory programming class. Nashwin Bharwani programmed the backend
program of the game, including but not limited to: move generation, game
rules, and board representations. Chen Liu focused on the front end user
interface of the game. David Xu programmed the AI system based off a
minimax algorithm with alpha beta pruning.